pyproject.toml
[tool.poetry]
name = "flask-taskq"
version = "0.1.1"
description = "Database driven task queue for Flask applications"
authors = ["Ivan Fedorov <inbox@titaniumhocker.ru>"]
license = "MIT"
repository = "https://github.com/TitaniumHocker/Flask-TaskQ"
documentation = "https://flask-taskq.rtfd.io"
keywords = ["flask", "extension", "task-schedules", "queue"]
readme = "README.rst"
packages = [{include = "flask_taskq"}]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Web Environment",
  "Framework :: Flask",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development :: Libraries"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/TitaniumHocker/Flask-TaskQ/issues"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
flask = ">=2.0.1,<3.0"
flask-sqlalchemy = ">=2.5.1"
sqlalchemy-utils = ">=0.38.2"
prettytable = ">=3.6.0,<4.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^5.5"}
pytest-cov = "^2.12.1"
flake8 = "<5.0.0"
flakeheaven = "<=3.0.0"
flake8-isort = "<6.0.0"
flake8-black = "<0.3.4"
flake8-docstrings = "<1.6.0"
flake8-bandit = "<4.0.0"
types-flask-sqlalchemy = "^2.5.9.3"
flake8-bugbear = "<=23.3.12"
sphinx = "<=6.1.3"
pallets-sphinx-themes = "^2.0.3"
tomlkit = "<=0.11.7"
sphinx-autodoc-typehints = "<=1.22"
flask-shell-ipython = "^0.5.1"

[tool.black]
line_length = 88

[tool.flakeheaven]
format = "colored"
max_line_length = 88
show_source = true

[tool.flakeheaven.plugins]
pyflakes = ["+*"]
pycodestyle = ["+*", "-W503", "-E501"]
flake8-docstrings = ["+*", "-D400", "-D107", "-D204", "-D105", "-D100"]
flake8-isort = ["+*"]
flake8-black = ["+*"]

[tool.flakeheaven.exceptions."tests/*"]
flake8-docstrings = ["-*"]
pycodestyle = ["-F401"]

[tool.isort]
profile = "black"
src_paths = ["flask_constance"]
line_length = 88

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = """
  --cov=flask_constance
"""

[tool.coverage.run]
source = [
    "flask_constance"
]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.mypy]
files="flask_constance"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
  "flask_admin.*",
  "flask_wtf.*",
  "wtforms.*",
]
ignore_missing_imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
